--- scripts/generate_manifest.py
@@ -1,229 +1,0 @@
-#!/usr/bin/env python3
-"""Generate or validate manifest.json from skill directories."""
-
-import argparse
-import json
-import re
-import sys
-from datetime import datetime, timezone
-from pathlib import Path
-
-
-def extract_version_from_skill(skill_path: Path) -> str:
-    """Extract version from SKILL.md frontmatter metadata."""
-    skill_md = skill_path / "SKILL.md"
-    if not skill_md.exists():
-        raise ValueError(f"SKILL.md not found in {skill_path}")
-
-    content = skill_md.read_text()
-
-    # parse YAML frontmatter
-    if not content.startswith("---"):
-        raise ValueError(f"SKILL.md in {skill_path} missing frontmatter")
-
-    end_idx = content.find("---", 3)
-    if end_idx == -1:
-        raise ValueError(f"SKILL.md in {skill_path} has unclosed frontmatter")
-
-    frontmatter = content[3:end_idx]
-
-    # extract version from metadata block
-    version_match = re.search(r'version:\s*["\']?([^"\'\n]+)["\']?', frontmatter)
-    if version_match:
-        return version_match.group(1).strip()
-
-    return "0.0.0"
-
-
-def get_skill_updated_at(skill_path: Path) -> str:
-    """Get the most recent modification time of any file in the skill directory."""
-    latest_mtime = 0.0
-    for file_path in skill_path.rglob("*"):
-        if file_path.is_file():
-            mtime = file_path.stat().st_mtime
-            if mtime > latest_mtime:
-                latest_mtime = mtime
-
-    if latest_mtime == 0.0:
-        return datetime.now(timezone.utc).strftime("%Y-%m-%dT%H:%M:%SZ")
-
-    return datetime.fromtimestamp(latest_mtime, timezone.utc).strftime(
-        "%Y-%m-%dT%H:%M:%SZ"
-    )
-
-
-# Repo-root files included in every skill's file list with an @root: prefix.
-# The CLI installer uses this prefix to fetch from the repo root instead of skills/<name>/.
-SHARED_ASSETS = [
-    "assets/databricks.svg",
-    "assets/databricks.png",
-]
-
-SKILL_METADATA = {
-    "databricks-core": {
-        "description": "Core Databricks skill for CLI, auth, and data exploration",
-        "experimental": False,
-    },
-    "databricks-apps": {
-        "description": "Databricks Apps development and deployment",
-        "experimental": False,
-    },
-    "databricks-jobs": {
-        "description": "Databricks Jobs orchestration and scheduling",
-        "experimental": False,
-    },
-    "databricks-lakebase": {
-        "description": "Databricks Lakebase database development",
-        "experimental": False,
-    },
-    "databricks-dabs": {
-        "description": "Declarative Automation Bundles (DABs) for deploying and managing Databricks resources",
-        "experimental": False,
-    },
-    "databricks-pipelines": {
-        "description": "Databricks Pipelines (DLT) for ETL and streaming",
-        "experimental": False,
-    },
-    "databricks-proto-first": {
-        "description": "Proto-first schema design for Databricks apps",
-        "experimental": False,
-    },
-}
-
-
-def generate_manifest(repo_root: Path) -> dict:
-    """Generate manifest from skill directories."""
-    # Load existing manifest to preserve base_revision fields
-    manifest_path = repo_root / "manifest.json"
-    existing_skills = {}
-    if manifest_path.exists():
-        existing_skills = json.loads(manifest_path.read_text()).get("skills", {})
-
-    for asset in SHARED_ASSETS:
-        if not (repo_root / asset).exists():
-            raise ValueError(f"Missing shared asset '{asset}' at repo root.")
-
-    skills = {}
-    skills_dir = repo_root / "skills"
-
-    for item in sorted(skills_dir.iterdir()):
-        if not item.is_dir():
-            continue
-        if item.name.startswith(".") or item.name == "scripts":
-            continue
-
-        skill_md = item / "SKILL.md"
-        if not skill_md.exists():
-            continue
-
-        files = sorted(
-            str(f.relative_to(item))
-            for f in item.rglob("*")
-            if f.is_file()
-        )
-        files += [f"@root:{f}" for f in SHARED_ASSETS]
-
-        if item.name not in SKILL_METADATA:
-            raise ValueError(f"Missing SKILL_METADATA entry for skill '{item.name}'. Add it to SKILL_METADATA dict.")
-
-        openai_yaml = item / "agents" / "openai.yaml"
-        if not openai_yaml.exists():
-            raise ValueError(f"Missing agents/openai.yaml in skill '{item.name}'. Each skill must include Codex marketplace metadata.")
-
-        metadata = SKILL_METADATA[item.name]
-        skill_entry = {
-            "version": extract_version_from_skill(item),
-            "description": metadata.get("description", ""),
-            "experimental": metadata.get("experimental", False),
-            "updated_at": get_skill_updated_at(item),
-            "files": files,
-        }
-
-        if metadata.get("min_cli_version"):
-            skill_entry["min_cli_version"] = metadata["min_cli_version"]
-
-        # Preserve base_revision from existing manifest
-        existing = existing_skills.get(item.name, {})
-        if "base_revision" in existing:
-            skill_entry["base_revision"] = existing["base_revision"]
-
-        skills[item.name] = skill_entry
-
-    return {
-        "version": "2",
-        "updated_at": datetime.now(timezone.utc).strftime("%Y-%m-%dT%H:%M:%SZ"),
-        "skills": skills,
-    }
-
-
-def normalize_manifest(manifest: dict) -> dict:
-    """Normalize manifest for comparison by excluding volatile fields."""
-    normalized = manifest.copy()
-    normalized.pop("updated_at", None)
-
-    skills = {}
-    for name, skill in manifest.get("skills", {}).items():
-        skill_copy = skill.copy()
-        skill_copy.pop("updated_at", None)
-        skill_copy.pop("base_revision", None)
-        skills[name] = skill_copy
-
-    normalized["skills"] = skills
-    return normalized
-
-
-def validate_manifest(repo_root: Path) -> bool:
-    """Validate that manifest.json is up to date. Returns True if valid."""
-    manifest_path = repo_root / "manifest.json"
-
-    if not manifest_path.exists():
-        print("ERROR: manifest.json does not exist", file=sys.stderr)
-        return False
-
-    current_manifest = json.loads(manifest_path.read_text())
-    expected_manifest = generate_manifest(repo_root)
-
-    # compare without timestamps
-    current_normalized = normalize_manifest(current_manifest)
-    expected_normalized = normalize_manifest(expected_manifest)
-
-    if current_normalized != expected_normalized:
-        print("ERROR: manifest.json is out of date", file=sys.stderr)
-        print("\nExpected:", file=sys.stderr)
-        print(json.dumps(expected_normalized, indent=2), file=sys.stderr)
-        print("\nActual:", file=sys.stderr)
-        print(json.dumps(current_normalized, indent=2), file=sys.stderr)
-        return False
-
-    print("manifest.json is up to date")
-    return True
-
-
-def main() -> None:
-    parser = argparse.ArgumentParser(description="Generate or validate manifest.json")
-    parser.add_argument(
-        "mode",
-        nargs="?",
-        default="generate",
-        choices=["generate", "validate"],
-        help="Mode: generate (creates manifest.json, default) or validate (checks if up to date)",
-    )
-
-    args = parser.parse_args()
-    repo_root = Path(__file__).parent.parent
-
-    match args.mode:
-        case "generate":
-            manifest = generate_manifest(repo_root)
-            manifest_path = repo_root / "manifest.json"
-            manifest_path.write_text(json.dumps(manifest, indent=2) + "\n")
-            print(f"Generated {manifest_path}")
-            print(f"Found {len(manifest['skills'])} skill(s): {', '.join(manifest['skills'].keys())}")
-
-        case "validate":
-            is_valid = validate_manifest(repo_root)
-            sys.exit(0 if is_valid else 1)
-
-
-if __name__ == "__main__":
-    main()
