--- skills/databricks/agents/openai.yaml
@@ -1,7 +1,0 @@
-interface:
-  display_name: "Databricks"
-  short_description: "CLI, auth, and data exploration"
-  icon_small: "./assets/databricks.svg"
-  icon_large: "./assets/databricks.png"
-  brand_color: "#FF3621"
-  default_prompt: "Use $databricks for Databricks CLI, auth, and data exploration."
